Separate bot and menu checker
